chore(weave): extract clickhouse/file_operations.py with CHInfraProtocol
Move file create, read, storage, and stats methods into FileOperationsMixin.
